+++ src/pools/selection-strategies/weighted-round-robin-worker-choice-strategy.ts
@@ -79,7 +79,7 @@ export class WeightedRoundRobinWorkerChoiceStrategy<
     const chosenWorkerNodeKey = this.nextWorkerNodeKey
     do {
       this.weightedRoundRobinNextWorkerNodeKey()
-    } while (!this.isWorkerNodeReady(this.nextWorkerNodeKey))
+    } while (!this.isWorkerNodeEligible(this.nextWorkerNodeKey))
     return chosenWorkerNodeKey
   }